Add require '...fixtures' to fix constant_table_saver dependency
When I removed the gem "constant_table_saver" from our gemfile the
our gemfile started to error with an "uninitialized constant error".

The error said that "ActiveRecord::FixtureSet" was not initialized.

I have added a require 'active_record/fixtures' directly in the class.
I don't know if this is a good/bad idea but it sovled the problem...
for now.